WIP: reverts all commits up to #968
