+++ .github/workflows/tests_windows.yml
@@ -31,7 +31,7 @@ jobs:
         run: |
           case "${TARGET}" in
             windows-amd64-unit-test-4-cpu)
-              CPU=4 TIMEOUT=50m make test
+              CPU=4 make test
               ;;
             *)
               echo "Failed to find target"
@@ -59,5 +59,3 @@ jobs:
         with:
           version: v2.1.6
       - run: make coverage
-        env:
-          TIMEOUT: 50m
