+++ .go-version
@@ -1,1 +1,1 @@
-1.24.5
+1.24.4
