+++ CHANGELOG/CHANGELOG-1.4.md
@@ -1,11 +1,10 @@
 
 <hr>
 
-## v1.4.2(2025-06-27)
+## v1.4.2(TBD)
 
 ### BoltDB
 - [Fix the compilation issue on aix, android and solaris due to wrong use of `maxMapSize`](https://github.com/etcd-io/bbolt/pull/990)
-- [Add protection on meta page when it's being written](https://github.com/etcd-io/bbolt/pull/1005)
 
 <hr>
